Introduce ByteBuffer-based NSPR PRFileDesc
Like BufferPRFD, ByteBufferPRFD implements a NSPR PRFileDesc layer.
Instead of maintaining a set of buffers up front, ByteBufferPRFD holds
a reference to a pair of j_bytebuffers, each of which in turn wraps a
single ByteBuffer instance. This allows for direct reading/writing of
the buffers passed to a SSLEngine's wrap/unwrap call without first
copying to an intermediate buffer.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>