Allow PK11Cert, PK11PrivKey to be duplicated
PK11Cert and PK11PrivKey are reference counted, so duplication (at the
NSS level) is of little work. Allowing duplication (at the Java level)
will allow JSSEngine to take ownership of its certs and keys, allowing
the original reference to go out of scope and be closed independently of
the lifecycle of the JSSEngine.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>